Add Update and Add DL logic

+++ src/main/java/com/JK/SIMS/models/IC_models/InventoryPageResponse.java
@@ -19,7 +19,7 @@ public class InventoryPageResponse {
 
     private Integer outgoingStockSize;
 
-    private Integer damageLossSize;
+    private Long damageLossSize;
 
     private PaginatedResponse<InventoryDataDTO> inventoryDataDTOList;
 }
